# contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: MaxMarket API
  version: "1.0.2"
  description: |
    MaxMarket B2B wholesale platform API.
    Base URL: /api/v1

    **Correlation ID:** Every request MAY include an `X-Correlation-ID` header (UUID).
    If provided, the server echoes it in the response. If absent, the server generates one.
    Every response carries `X-Correlation-ID` in the response header.
    Every error response includes `correlationId` in the body.

    **Error envelope:** All error responses use the `ErrorResponse` schema.

    **Public catalog:** `GET /catalog/products` and `GET /catalog/products/{id}` require no
    authentication and return `PublicProduct` — price fields are entirely omitted.
    Authenticated requests to those same endpoints return `Product` with role-appropriate
    price fields resolved server-side.

    **Optimistic locking:** `POST /orders/{id}/approve` AND `PUT /orders/{id}` (when editing
    an Approved order) both require the current `versionLock` value to prevent concurrent
    conflicting writes.

    **Draft order pricing:** At Draft creation, `groupDiscount = 0` and `finalPrice = basePrice`
    (captured from current `pricePerUnit`). Both values are recalculated on submission from
    the then-current client group discount rule (DN-PRICE-02).

    **Decision Needed (open):**
    - DA-05: Can an Admin change a variant's SKU if non-Draft orders reference it?
...
      Current spec assumes URL-only (CDN pre-upload pattern).

...

...
servers:
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── HEADERS ────────────────────────────────────────────────────────────────
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    CorrelationIdRequestHeader:
      name: X-Correlation-ID
      in: header
      required: false
      description: Optional client-provided correlation UUID. Server echoes it in the response header.
      schema:
        type: string
        format: uuid

  # ─── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    OrderStatus:
      type: string
      enum: [draft, submitted, approved, fulfilled, rejected, cancelled, returned]
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      description: |
        Roles: super_admin, admin, manager (all orders); agent (own orders only, scoped by agent_id);
        client (own orders only, scoped by client_id).
        Scoping enforced server-side regardless of query filters provided.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: clientId
          in: query
          schema:
            type: string
            format: uuid
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated orders
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Orders]
      summary: Create draft order
      description: |
        Roles: agent only. The client must be in the agent's assigned client list.
        Prices are NOT set by the caller — they are computed server-side from product_variants
        and client_groups on submission. warehouseId defaults to the single v1 warehouse if omitted.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, lineItems]
              properties:
                clientId:
                  type: string
                  format: uuid
                lineItems:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [variantId, qty]
                    properties:
                      variantId:
                        type: string
                        format: uuid
                      qty:
                        type: integer
                        minimum: 1
                      warehouseId:
                        type: string
                        format: uuid
                        description: Defaults to the default v1 warehouse if omitted.
                notes:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Draft order created
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Client not assigned to this agent
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (e.g. qty below min_order_qty)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      description: |
        All authenticated roles. Scoping enforced server-side:
        agents see own orders; clients see own orders; managers/admins see all.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Order
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (order belongs to another agent/client)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Orders]
      summary: Edit order
      description: |
        **Agent editing Draft order:** Roles: agent (own drafts only). No versionLock required.

        **Admin editing Approved order (creates version):**
        Roles: super_admin, admin only. `versionLock` REQUIRED to prevent concurrent version
        creation conflicts (mandatory standard: optimistic locking on version edit).
        Editing an Approved order:
        1. Creates an immutable snapshot in order_versions
        2. Resets order status to `submitted`
        3. Increments currentVersion and versionLock
        4. Requires re-approval (stock delta checked at re-approval, not at edit — DN-STK-02)

        Attempting to edit a Submitted or Rejected order by any role returns ORDER_NOT_EDITABLE.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lineItems:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [variantId, qty]
                    properties:
                      variantId:
                        type: string
                        format: uuid
                      qty:
                        type: integer
                        minimum: 1
                      warehouseId:
                        type: string
                        format: uuid
                notes:
                  type: string
                  nullable: true
                versionLock:
                  type: integer
                  description: |
                    REQUIRED when editing an Approved order (Admin/Super Admin only).
                    Must match current versionLock from GET /orders/{id}.
                    Ignored (and must be omitted) for Draft order edits by Agent.
      responses:
        '200':
          description: |
            Updated order. For Admin Approved edits, status will be `submitted`,
            currentVersion incremented, and versionLock incremented.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (wrong role or not owner)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            Optimistic lock conflict (errorCode: OPTIMISTIC_LOCK_CONFLICT).
            Stale versionLock — another Admin edit occurred concurrently. Re-fetch and retry.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Order not in editable state (errorCode: ORDER_NOT_EDITABLE) or validation error.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Orders]
      summary: Delete draft order (soft delete)
      description: "Roles: agent (own Draft orders only). Non-Draft orders cannot be deleted."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '204':
          description: Deleted (no content)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
        '403':
          description: Forbidden
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order is not in Draft status
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/submit:
    post:
      tags: [Orders]
      summary: Submit draft order for approval
      description: |
        Roles: agent (own Draft orders only).
        Recalculates all line item prices from the current client group discount (DN-PRICE-02).
        Transitions order from Draft → Submitted.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Order submitted. Response includes recalculated finalPrices.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Not in Draft status or validation error
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/approve:
    post:
      tags: [Orders]
      summary: Approve submitted order
      description: |
        Roles: manager only.

        Stock check (PRD §4.3, FS-ORD-01): Approval is BLOCKED if any line item has
        `available_qty < order_qty`. Error code INSUFFICIENT_STOCK with per-line details.

        Optimistic locking (EC-ORD-05): `versionLock` from GET /orders/{id} MUST be sent.
        If another manager approved concurrently, returns 409 OPTIMISTIC_LOCK_CONFLICT.

        On success: reserved_qty += order_qty for all line items. Status → Approved.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [versionLock]
              properties:
                versionLock:
                  type: integer
                  description: |
                    Current versionLock value from GET /orders/{id}.
                    Prevents dual-approval race condition (EC-ORD-05).
      responses:
        '200':
          description: Approved. Stock reserved.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (manager role required)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Optimistic lock conflict (errorCode OPTIMISTIC_LOCK_CONFLICT)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Insufficient stock (errorCode INSUFFICIENT_STOCK) — details array lists each
            failing line item per PRD FS-ORD-01.
            OR wrong order state (errorCode ORDER_NOT_EDITABLE).
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - type: object
                    properties:
                      details:
                        type: array
                        items:
                          $ref: '#/components/schemas/InsufficientStockDetail'
              examples:
                insufficient_stock:
                  summary: Insufficient stock error
                  value:
                    errorCode: INSUFFICIENT_STOCK
                    message: "Approval blocked: insufficient stock for 2 line items."
                    details:
                      - lineItemId: "a1b2c3d4-..."
                        variantId: "e5f6g7h8-..."
                        sku: "SKU-001"
                        requestedQty: 50
                        availableQty: 30
                        reservedQty: 10
                    correlationId: "550e8400-..."
                wrong_state:
                  summary: Wrong order state
                  value:
                    errorCode: ORDER_NOT_EDITABLE
                    message: "Order is not in Submitted state."
                    details: null
                    correlationId: "550e8400-..."
  /orders/{id}/reject:
    post:
      tags: [Orders]
      summary: Reject submitted order
      description: |
        Roles: manager only.
        Rejection reason is optional (DN-ORD-01). Stored in audit log.
        Rejected orders are terminal — agent must create a new Draft to retry.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  nullable: true
                  description: Optional rejection reason. Stored in audit.
      responses:
        '200':
          description: Rejected
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (manager role required)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order is not in Submitted state
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/fulfill:
    post:
      tags: [Orders]
      summary: Mark order as fulfilled
      description: |
        Roles: manager only.
        Transitions Approved → Fulfilled.
        Stock effect: reserved_qty -= order_qty AND available_qty -= order_qty.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Fulfilled. Stock decremented.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (manager role required)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order is not in Approved state
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel approved order
      description: |
        Roles: super_admin, admin, manager (DN-ORD-02).
        Only Approved orders can be cancelled. Terminal state.
        Stock effect: reserved_qty -= order_qty (reservation released).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdRequestHeader'
      responses:
        '200':
          description: Cancelled. Reserved stock released.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden (requires super_admin, admin, or manager role)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order is not in Approved state
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...